create runtime/survey_test.exs and split test between it and broker_test (best-effort)
